+++ internal/provider/provider.go
@@ -174,6 +174,7 @@ func (p *SonatypeIqProvider) Resources(ctx context.Context) []func() resource.Re
 		organization.NewApplicationCategoryResource,
 		organization.NewOrganizationResource,
 		organization.NewOrganizationRoleMembershipResource,
+		role.NewRoleResource,
 		scm.NewSourceControlResource,
 		system.NewConfigCrowdResource,
 		system.NewConfigMailResource,