+++ CHANGELOG.md
@@ -9,6 +9,10 @@ See the [Migration Guide](./MIGRATION-1.0.md) for help migrating from v0.x.x ver
 * Updated Data Sources to always have an `id` and be defined consistently [GH-63]. See [Migration Guide](./MIGRATION-1.0.md) for impacted data sources.
 * Updated Resources to ensure they always have an `id` and `last_updated` for consistency [Gh-64]. See [Migration Guide](./MIGRATION-1.0.md) for impacted resources.
 
+FEATURES:
+
+* **New Resource:** `sonatypeiq_role` [GH-16]
+
 ENHANCEMENTS:
 
 * Adopted shared library to improve mantainability and consistency [GH-60]